refactor: remove supabase provider and listener
They look redundant so avoid it.
